% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{create_kernel_functions}
\alias{create_kernel_functions}
\title{Create kernel functions based on configuration}
\usage{
create_kernel_functions(
  kernel.fun = "Schennach2004",
  if_approx_kernel = TRUE,
  kernel.resol = 1000
)
}
\arguments{
\item{kernel.fun}{A string specifying the kernel function to be used.}

\item{if_approx_kernel}{Logical. If TRUE, uses approximations for the kernel function.}

\item{kernel.resol}{The resolution for kernel function approximation.}
}
\value{
A list containing kernel function, its Fourier transform, and the kernel type
}
\description{
Create kernel functions based on configuration
}
