% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.has_lp_solver}
\alias{.has_lp_solver}
\title{Check if LP solver is available}
\usage{
.has_lp_solver()
}
\value{
Logical indicating if Rglpk is available
}
\description{
Check if LP solver is available
}
\keyword{internal}
