% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.prepare_lp_problem}
\alias{.prepare_lp_problem}
\title{Prepare LP problem matrices for envelope optimization}
\usage{
.prepare_lp_problem(ln_xi_range, avg_phi_log_values, r_max = 50)
}
\arguments{
\item{ln_xi_range}{Numeric vector of log-frequencies}

\item{avg_phi_log_values}{Numeric vector of log-Fourier-transform values}

\item{r_max}{Maximum allowed value for r parameter (default 50)}
}
\value{
List with components: obj (objective coefficients),
mat (constraint matrix), rhs (RHS vector), bounds (variable bounds)
}
\description{
Prepare LP problem matrices for envelope optimization
}
\keyword{internal}
