% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.solve_lp_envelope}
\alias{.solve_lp_envelope}
\title{Solve LP problem for envelope optimization}
\usage{
.solve_lp_envelope(lp_problem)
}
\arguments{
\item{lp_problem}{List from .prepare_lp_problem()}
}
\value{
List with status, solution vector, and objective value
}
\description{
Solve LP problem for envelope optimization
}
\keyword{internal}
