% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_est_byx}
\alias{get_est_byx}
\title{Estimation of bias byx}
\usage{
get_est_byx(Y, X, ...)
}
\arguments{
\item{Y}{A numerical vector representing the sample data of variable Y.}

\item{X}{A numerical vector representing the sample data of variable X.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
A scalar representing the bias byx estimate.
}
\description{
Estimation of bias byx
}
\keyword{internal}
