% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sigma}
\alias{get_sigma}
\title{Estimation of sigma}
\usage{
get_sigma(X, x, h, inf_k)
}
\arguments{
\item{X}{A numerical vector of sample data.}

\item{x}{A scalar representing the point where the density is estimated.}

\item{h}{A scalar bandwidth parameter.}

\item{inf_k}{Kernel function used for the computation.}
}
\value{
A scalar representing the sigma estimate at point x. Returns 0 if the density estimate is negative.
}
\description{
Computes the sigma estimate for given parameters.
}
\keyword{internal}
