% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sigma_yx}
\alias{get_sigma_yx}
\title{Estimation of sigma_yx}
\usage{
get_sigma_yx(Y, X, x, h, inf_k)
}
\arguments{
\item{Y}{A numerical vector representing the sample data of variable Y.}

\item{X}{A numerical vector representing the sample data of variable X.}

\item{x}{The specific point at which sigma_yx is to be estimated.}

\item{h}{A bandwidth parameter used in the kernel function for smoothing.}

\item{inf_k}{A kernel function used to weigh observations in the neighborhood of point x.}
}
\value{
Returns a scalar representing the estimated value of sigma_yx at the point x.
Returns 0 if either fyx or conditional variance is negative.
}
\description{
Estimation of sigma_yx
}
\keyword{internal}
