% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kernel_reg}
\alias{kernel_reg}
\title{Kernel Regression function}
\usage{
kernel_reg(X, Y, x, h, kernel_func)
}
\arguments{
\item{X}{A numerical vector representing the sample data of variable X.}

\item{Y}{A numerical vector representing the sample data of variable Y.}

\item{x}{The point at which the regression function is to be estimated.}

\item{h}{A bandwidth parameter that determines the weight assigned to each observation in X.}

\item{kernel_func}{A function that computes the weight of each observation based on its distance to x.}
}
\value{
Returns a scalar representing the estimated value of the regression function at the point x.
}
\description{
Kernel Regression function
}
\keyword{internal}
