% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{normal_kernel_ft}
\alias{normal_kernel_ft}
\title{Fourier Transform of Normal Kernel}
\usage{
normal_kernel_ft(xi)
}
\arguments{
\item{xi}{A numerical value or vector representing the frequency domain.}
}
\value{
Returns the value of the Fourier transform of the Normal kernel at the given frequency/frequencies.
}
\description{
Fourier Transform of Normal Kernel
}
\keyword{internal}
