% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\name{plot_regression_internal}
\alias{plot_regression_internal}
\title{Internal Helper: Plot Conditional Expectation}
\usage{
plot_regression_internal(x, fill_ci, alpha_ci, point_alpha, point_color)
}
\arguments{
\item{x}{An object of class bbnp_regression}

\item{fill_ci}{Color for confidence interval ribbon}

\item{alpha_ci}{Transparency for confidence interval ribbon}

\item{point_alpha}{Transparency for data points}

\item{point_color}{Color for data points}
}
\value{
A ggplot2 object
}
\description{
Creates regression plot with confidence interval and data points
}
\keyword{internal}
