% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-print.R
\name{print.bbnp_density}
\alias{print.bbnp_density}
\title{Print Method for bbnp_density Objects}
\usage{
\method{print}{bbnp_density}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class bbnp_density}

\item{digits}{Number of digits to display (default: 4)}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for bbnp_density Objects
}
