% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-print.R
\name{print.bbnp_regression}
\alias{print.bbnp_regression}
\title{Print Method for bbnp_regression Objects}
\usage{
\method{print}{bbnp_regression}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class bbnp_regression}

\item{digits}{Number of digits to display (default: 4)}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for bbnp_regression Objects
}
