% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sample.R
\name{rpoly01}
\alias{rpoly01}
\title{Generate n samples from the distribution}
\usage{
rpoly01(n, k = 5)
}
\arguments{
\item{n}{The number of samples to generate.}

\item{k}{The exponent in the distribution function, defaults to 5.}
}
\value{
A vector of \code{n} samples from the specified polynomial distribution.

CDF: f(x) = (x-1)^k + 1
}
\description{
Generate n samples from the distribution
}
\keyword{internal}
