% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sample.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample Data}
\format{
A data frame with 1000 rows and 2 variables:
\describe{
\item{X}{Numeric vector, generated from 2 fold uniform distribution.}
\item{Y}{Numeric vector, \code{Y = -X^2 + 3*X + rnorm(1000)*X}.}
}
}
\usage{
sample_data
}
\description{
Sample Data
}
\keyword{datasets}
