% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{sinc}
\alias{sinc}
\title{Infinite Kernel Function}
\usage{
sinc(u)
}
\arguments{
\item{u}{A numerical value or vector where the sinc function is evaluated.}
}
\value{
The value of the sinc function at each point in \code{u}.
}
\description{
Infinite Kernel Function
}
\keyword{internal}
