% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{sinc_ft}
\alias{sinc_ft}
\title{Define the closed form FT of the infinite order kernel sin(x)/(pi*x)}
\usage{
sinc_ft(x)
}
\arguments{
\item{x}{A numerical value or vector where the Fourier Transform is evaluated.}
}
\value{
The value of the Fourier Transform of the sinc function at each point in \code{x}.
}
\description{
Define the closed form FT of the infinite order kernel sin(x)/(pi*x)
}
\keyword{internal}
