% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-summary.R
\name{summary.bbnp_density}
\alias{summary.bbnp_density}
\title{Summary Method for bbnp_density Objects}
\usage{
\method{summary}{bbnp_density}(object, ...)
}
\arguments{
\item{object}{An object of class bbnp_density}

\item{...}{Additional arguments (unused)}
}
\value{
An object of class summary.bbnp_density
}
\description{
Summary Method for bbnp_density Objects
}
