% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-summary.R
\name{summary.bbnp_regression}
\alias{summary.bbnp_regression}
\title{Summary Method for bbnp_regression Objects}
\usage{
\method{summary}{bbnp_regression}(object, ...)
}
\arguments{
\item{object}{An object of class bbnp_regression}

\item{...}{Additional arguments (unused)}
}
\value{
An object of class summary.bbnp_regression
}
\description{
Summary Method for bbnp_regression Objects
}
