% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sample.R
\name{true_density_2fold}
\alias{true_density_2fold}
\title{True density of 2-fold uniform distribution}
\usage{
true_density_2fold(x)
}
\arguments{
\item{x}{A numerical value or vector where the true density function is evaluated.}
}
\value{
The value of the true density of the 2-fold uniform distribution at each point in \code{x}.
}
\description{
True density of 2-fold uniform distribution
}
\keyword{internal}
