% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{DATA_PATH}
\alias{DATA_PATH}
\title{The Path to the Data Folder}
\value{
The path to the package's internal data folder as a character string.
}
\description{
This variable provides the path to the \code{data} folder within the package.
}
