% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{epanechnikov_kernel}
\alias{epanechnikov_kernel}
\title{Epanechnikov Kernel}
\usage{
epanechnikov_kernel(u)
}
\arguments{
\item{u}{A numerical value or vector representing the input to the kernel function.}
}
\value{
Returns the value of the Epanechnikov kernel function at the given input.
}
\description{
Epanechnikov Kernel
}
