% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{epanechnikov_kernel_ft}
\alias{epanechnikov_kernel_ft}
\title{Fourier Transform Epanechnikov Kernel}
\usage{
epanechnikov_kernel_ft(xi)
}
\arguments{
\item{xi}{A numerical value or vector representing the frequency domain.}
}
\value{
Returns the value of the Fourier transform of the Epanechnikov kernel at the given frequency/frequencies.
}
\description{
Fourier Transform Epanechnikov Kernel
}
