% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{EXT_DATA_PATH}
\alias{EXT_DATA_PATH}
\title{The Path to the External Data Folder for Non-R Data Files}
\value{
The path to the package's external data folder (for non-standard R data files) as a character string.
}
\description{
This variable provides the path to the \code{extdata} folder within the package,
where non-standard R data files are stored.
}
