% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_avg_f1x}
\alias{get_avg_f1x}
\title{Kernel point estimation}
\usage{
get_avg_f1x(X, x, h, inf_k)
}
\arguments{
\item{X}{A numerical vector of sample data.}

\item{x}{A scalar representing the point where the density is estimated.}

\item{h}{A scalar bandwidth parameter.}

\item{inf_k}{Kernel function used for the computation.}
}
\value{
A scalar representing the kernel density estimate at point x.
}
\description{
Computes the point estimate using the specified kernel function.
}
