% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_avg_phi_log}
\alias{get_avg_phi_log}
\title{Compute log sample average of fourier transform and get mod}
\usage{
get_avg_phi_log(Y = 1, X, ln_xi)
}
\arguments{
\item{Y}{A numerical vector representing the sample data of variable Y.}

\item{X}{A numerical vector representing the sample data of variable X.}

\item{ln_xi}{A single numerical value representing the log frequency at which the Fourier transform
is computed.}
}
\value{
Returns the log sample estimation of expected Fourier transform at frequency \code{xi}.
}
\description{
Compute log sample average of fourier transform and get mod
}
