% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_conditional_var}
\alias{get_conditional_var}
\title{get the conditional variance of Y on X for given x}
\usage{
get_conditional_var(X, Y, x, h, kernel_func)
}
\arguments{
\item{X}{A numerical vector representing the sample data of variable X.}

\item{Y}{A numerical vector representing the sample data of variable Y.}

\item{x}{The specific point at which the conditional variance is to be calculated.}

\item{h}{A bandwidth parameter used in the kernel function for smoothing.}

\item{kernel_func}{A kernel function used to weigh observations in the neighborhood of point x.}
}
\value{
Returns a non-negative scalar representing the estimated conditional variance of Y given X at the point x.
Returns 0 if the computed variance is negative.
}
\description{
get the conditional variance of Y on X for given x
}
