% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_est_B}
\alias{get_est_B}
\title{get the estimation of B}
\usage{
get_est_B(Y)
}
\arguments{
\item{Y}{A numerical vector representing the sample data of variable Y.}
}
\value{
The mean of the absolute values of the elements in Y, representing the estimated value of \eqn{B}.
}
\description{
get the estimation of B
}
