% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_est_b1x}
\alias{get_est_b1x}
\title{Estimation of bias b1x}
\usage{
get_est_b1x(X, ...)
}
\arguments{
\item{X}{A numerical vector representing the sample data of variable X.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
A scalar representing the bias b1x estimate.
}
\description{
Computes the bias estimate for given parameters.
}
