% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_est_vy}
\alias{get_est_vy}
\title{get the estimation of Vy}
\usage{
get_est_vy(Y)
}
\arguments{
\item{Y}{A numerical vector representing the sample data of variable Y.}
}
\description{
get the estimation of Vy
}
