% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_xi_interval}
\alias{get_xi_interval}
\title{get xi interval}
\usage{
get_xi_interval(Y = 1, X, methods = "Schennach")
}
\arguments{
\item{Y}{A numerical vector representing the sample data of variable Y.}

\item{X}{A numerical vector representing the sample data of variable X.}

\item{methods}{A character string indicating the method to use for calculating the
xi interval. Supported methods are "Schennach" and "Schennach_loose".
Defaults to "Schennach".}
}
\value{
A list containing the lower (\code{xi_lb}) and upper (\code{xi_ub}) bounds of the xi interval.
}
\description{
get xi interval
}
\details{
The "Schennach" method computes the xi interval by performing a test based on the
Schennach's theorem, adjusting the upper bound \code{xi_ub} if the test condition is met.
The "Schennach_loose" method provides a looser calculation of the xi interval without
performing the Schennach's test.
}
