% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{normal_kernel}
\alias{normal_kernel}
\title{Normal Kernel Function}
\usage{
normal_kernel(u)
}
\arguments{
\item{u}{A numerical value or vector representing the input to the kernel function.}
}
\value{
Returns the value of the Normal kernel function at the given input.
}
\description{
Normal Kernel Function
}
