% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{W_kernel}
\alias{W_kernel}
\title{Define the inverse Fourier transform function of W}
\usage{
W_kernel(u, L = 10)
}
\arguments{
\item{u}{A numerical value or vector representing the time or space domain.}

\item{L}{The limit for numerical integration, defines the range of integration as \eqn{[-L, L]}.
Defaults to 10.}
}
\value{
A numerical value or vector representing the inverse Fourier transform of the infinite
order kernel at the given time or space point(s).
}
\description{
Define the inverse Fourier transform function of W
}
